rag/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
pytest==7.4.3
pyjwt==2.8.0
requests==2.31.0

# PDF Processing
PyMuPDF==1.23.14
pdfplumber==0.10.3
pytesseract==0.3.10
Pillow==10.1.0

# NLP and Embeddings
sentence-transformers==2.2.2
transformers==4.35.2
torch==2.1.1
numpy==1.24.3

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# OCR and Image Processing
opencv-python==4.8.1.78
easyocr==1.7.0

# Math Formula Processing
sympy==1.12